keep refactorting, fixed the lightning bug
