+++ readme.md
@@ -1,9 +1,7 @@
 # FEATURES
-- [ ] Normalize the movement vector for consistent speed in all directions, including diagonals.
 - [ ] Add cub3d.
-- [ ] Implement a digit scroll animation effect for the divergence meter to enhance visual feedback and user experience.
-- [ ] Store the divergence meter value in local storage to persist the state across sessions.
-- Use original Font instead of image for divergence meter.
+- [ ] Add an interactive resume.
+- [ ] Add an earthquake when microwave is activated.
 
 # BUGS
 
@@ -14,6 +12,11 @@
 # OPTIMIZATION
 - [ ] Remove event listeners when they are no longer needed to prevent memory leaks and improve performance. (you can use  { once: true })
 
-
 # MAINTAINABILITY
 - [ ] Reorganize assets in the public folders for better structure, accessibility, and maintainability.
+
+# IMPROVEMENT
+- [ ] Use original Font instead of image for divergence meter.
+- [ ] Store the divergence meter value in local storage to persist the state across sessions.
+- [ ] Normalize the movement vector for consistent speed in all directions, including diagonals.
+- [ ] Customize Achievement banner.
