+++ src/dialogue.js
@@ -1,3 +1,5 @@
+import { ensureCanvasFocus } from "./utils";
+
 export function displayDialogue(text, onDisplayEnd) {
     const dialogueUI = document.getElementById("textbox-container");
     const dialogue = document.getElementById("dialogue");
